module + hash
